% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsLayer.R
\docType{methods}
\name{getData,qgsLayer-method}
\alias{getData,qgsLayer-method}
\title{Read the data from a layer to R.}
\usage{
\S4method{getData}{qgsLayer}(object)
}
\arguments{
\item{object}{A qgsLayer object.}
}
\description{
Read the data from a layer to R.
}
