% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsProject.R
\docType{methods}
\name{openLayer,qgsProject-method}
\alias{openLayer,qgsProject-method}
\title{Open a given layer within a qgs project.}
\usage{
\S4method{openLayer}{qgsProject}(object, name)
}
\arguments{
\item{object}{a qgsProject object.}

\item{name}{Name of the layer to be opened.}
}
\description{
Open a given layer within a qgs project.
}
