% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsLayer.R
\docType{methods}
\name{plot,qgsLayer,ANY-method}
\alias{plot,qgsLayer,ANY-method}
\title{Plot a QGIS layer}
\usage{
\S4method{plot}{qgsLayer,ANY}(x, y, ...)
}
\arguments{
\item{x}{A qgsLayer object.}

\item{y}{Ignored argument.}

\item{...}{Argument passed to sp::plot.}
}
\description{
Read data from a QGIS layer and plot it.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
