% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsLayer.R
\docType{class}
\name{qgsLayer-class}
\alias{qgsLayer-class}
\alias{qgsLayer}
\title{A layer of a QGIS project}
\description{
This class represents a layer of a QGIS project.
Please do not create it manually. Instead, use \code{\link{openLayer,qgsProject-method}}.
}
\section{Slots}{

\describe{
\item{\code{project}}{A qgsProject.}

\item{\code{name}}{The name of the layer.}

\item{\code{provider}}{Name of its provider (WMS, ogr, gdal).}

\item{\code{source}}{Description of where the data is stored.}
}}

