% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsProject.R
\docType{class}
\name{qgsProject-class}
\alias{qgsProject-class}
\alias{qgsProject}
\title{An S4 class to represent a QGIS project file.}
\description{
An S4 class to represent a QGIS project file.
}
\section{Slots}{

\describe{
\item{\code{file}}{A character with the file.}

\item{\code{xml}}{The xml content of the file decoded as a list.}
}}

