% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsLayer.R
\docType{methods}
\name{readData,qgsLayer-method}
\alias{readData,qgsLayer-method}
\alias{readData}
\title{Read data from a QGIS layer}
\usage{
\S4method{readData}{qgsLayer}(object)
}
\arguments{
\item{object}{A qgsLayer object.}
}
\description{
Read data from the data sorce pointed by a QGIS layer into an sp object. It automatically recognizes the type of the data source in order to read it.
}
\seealso{
\code{\link{readData,qgsLayer-method}}
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
