% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsLayer.R
\docType{methods}
\name{readDataCode,qgsLayer-method}
\alias{readDataCode,qgsLayer-method}
\alias{readDataCode}
\title{Code to read data}
\usage{
\S4method{readDataCode}{qgsLayer}(object)
}
\arguments{
\item{object}{A qgsLayer object.}
}
\description{
Return a string with R code to the data from a layer using the data source directly.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
