% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsLayer.R
\docType{methods}
\name{show,qgsLayer-method}
\alias{show,qgsLayer-method}
\title{Show the properties of a QGIS layer}
\usage{
\S4method{show}{qgsLayer}(object)
}
\arguments{
\item{object}{A qgsLayer object.}
}
\description{
Show the following properties of a QGIS layer: name, provider, and source.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
