% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsProject.R
\docType{methods}
\name{show,qgsProject-method}
\alias{show,qgsProject-method}
\title{Show the properties of a QGIS project: file and layers.}
\usage{
\S4method{show}{qgsProject}(object)
}
\arguments{
\item{object}{A qgsProject object.}
}
\description{
Show the properties of a QGIS project: file and layers.
}
\author{
Pedro R. Andrade, \email{pedro.andrade@inpe.br}
}
